\name{dummy-package}
\alias{dummy-package}
\alias{dummy}
\docType{package}
\title{
\packageTitle{dummy}
}
\description{
\packageDescription{dummy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dummy}
\packageIndices{dummy}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dummy}

Maintainer: \packageMaintainer{dummy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
