% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_function}
\alias{dummy_function}
\title{A function that returns TRUE}
\usage{
dummy_function(x, one_argument, another_argument)
}
\arguments{
\item{x}{first argument. Ignored.}

\item{one_argument}{second argument. Also ignored.}

\item{another_argument}{third arument. You guessed it: Ignored.}
}
\description{
This function does nothing. It merely returns \code{TRUE}.
}
\details{
Nothing to see here, move along.
}
\examples{
# This example shows how useless this function is
dummy_function()

}
