% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_function2}
\alias{dummy_function2}
\title{Another useless function}
\usage{
dummy_function2(x)
}
\arguments{
\item{x}{it's the only parameter and it's ignored.}
}
\description{
This function returns \code{FALSE}.
}
\details{
Why would you use this function instead of just typing \code{FALSE}?
}
\examples{
# Proof that it works
FALSE == dummy_function2()

}
