% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dummyNews.R
\name{dummyNews}
\alias{dummyNews}
\title{Display the NEWS file}
\usage{
dummyNews()
}
\description{
\code{dummyNews} shows the NEWS file of the \code{dummy} package.
}
\examples{
dummyNews()
}
\author{
Authors: Michel Ballings and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}

