% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_numbers.R
\docType{data}
\name{random_numbers}
\alias{random_numbers}
\title{Random Numbers Generator}
\format{
An object of class \code{numeric} of length 10.
}
\source{
Coronavirus package \url{https://panagiotis-stylianos.netlify.app}{Web}
}
\usage{
Random Numbers for random tests
}
\description{
This function produces 10 random numbers.
}
\details{
More specifically the numbers come from the standard normal distribution.
}
\keyword{datasets}
