\name{test}
\alias{test}
\title{Test}
\usage{
  test(x = "test")
}
\arguments{
  \item{x}{An object}
}
\description{
  Test 2
}
\section{Suggestion}{
  Looky here: \enumerate{ \item{\bold{Item 1} - Foo}
  \item{\bold{Item 2} - Bar:
  \href{https://github.com/hakimel/reveal.js/\#installation}{click
  me}} }
}

