\name{fun1}
\alias{fun1}
\alias{fun2}
\alias{fun3}
\title{Readabiltiy stat}
\usage{
  fun1(X, ...)

  fun2(X, simplify = FALSE, ...)

  fun3(X, noargs = TRUE, ...)
}
\arguments{
  \item{X}{Something called X}

  \item{...}{those crazy dots}

  \item{simplify}{shut yo face}

  \item{noargs}{no arguments}
}
\description{
  fun1 - blah 2

  fun2 - blah2

  fun3 - blah 3
}

