% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy-package.R
\docType{package}
\name{dummy-package}
\alias{dummy}
\alias{dummy-package}
\title{dummy: This packages is a dummy package to help me learn the structure of R packages}
\description{
I created this dummy package for the purpose of learning the structure and the workflow of creating R packages in order to create more advanced packages in the coming future.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petestylianos/dummy}
  \item Report bugs at \url{https://github.com/petestylianos/dummy/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
