% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dummy_function.R
\name{dummy_function}
\alias{dummy_function}
\title{A dummy function}
\usage{
dummy_function(dummy = TRUE)
}
\arguments{
\item{dummy}{Is the function a dummy function? Defaults to TRUE.}
}
\description{
This function is a dummy function.
}
\examples{
dummy_function()
}
\keyword{dummy}
\keyword{function}

