% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_function.R
\name{dummy_function}
\alias{dummy_function}
\title{A dummy function}
\usage{
dummy_function(x)
}
\arguments{
\item{x}{a numeric input}
}
\value{
numeric value
}
\description{
I am serious this is very dum.
}
\examples{
dummy_function(3)


}
