% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dummyNews}
\alias{dummyNews}
\title{Display the NEWS file}
\usage{
dummyNews()
}
\description{
\code{dummyNews} shows the NEWS file of the \code{dummy} package.
}
\examples{
dummyNews()
}
\author{
Authors: Michel Ballings and Dirk Van den Poel, Maintainer: \email{Michel.Ballings@GMail.com}
}

